// apps/storefront/prismicio-types.d.ts
export interface SwiperSliceProductFeatureItem {
    /**
     * Desktop Image field in *Swiper → Items*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: swiper.items[].desktop_image
     * - **Documentation**: https://prismic.io/docs/field#image
     */
    desktop_image: prismic.ImageField<never>;

    /**
     * Mobile Image field in *Swiper → Items*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: swiper.items[].mobile_image
     * - **Documentation**: https://prismic.io/docs/field#image
     */
    mobile_image: prismic.ImageField<never>;

    /**
     * Description field in *Swiper → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: swiper.items[].description
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    description: prismic.KeyTextField;
}

/**
 * Product Feature variation for Swiper Slice
 *
 * - **API ID**: `product_feature`
 * - **Description**: Default
 * - **Documentation**: https://prismic.io/docs/slice
 */
export type SwiperSliceProductFeature = prismic.SharedSliceVariation<
    'product_feature',
    Simplify<SwiperSliceProductFeaturePrimary>,
    Simplify<SwiperSliceProductFeatureItem>
>;
// ...
export type SwiperSliceProductShopify = prismic.SharedSliceVariation<
    'product_shopify',
    Simplify<SwiperSliceProductShopifyPrimary>,
    never
>;
// ...
export interface SwiperSliceCollectionItem {
    /**
     * Desktop Image field in *Swiper → Items*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: swiper.items[].desktop_image
     * - **Documentation**: https://prismic.io/docs/field#image
     */
    desktop_image: prismic.ImageField<never>;

    /**
     * Mobile Image field in *Swiper → Items*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: swiper.items[].mobile_image
     * - **Documentation**: https://prismic.io/docs/field#image
     */
    mobile_image: prismic.ImageField<never>;

    /**
     * Context field in *Swiper → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: Context
     * - **API ID Path**: swiper.items[].context
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    context: prismic.KeyTextField;

    /**
     * Title field in *Swiper → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: MAAP X 100% EYEWEAR
     * - **API ID Path**: swiper.items[].title
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    title: prismic.KeyTextField;

    /**
     * Button Label field in *Swiper → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: Discover
     * - **API ID Path**: swiper.items[].button_label
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    button_label: prismic.KeyTextField;

    /**
     * Link field in *Swiper → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: /collections/new-arrivals
     * - **API ID Path**: swiper.items[].link
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    link: prismic.KeyTextField;

    /**
     * Button Variant field in *Swiper → Items*
     *
     * - **Field Type**: Select
     * - **Placeholder**: *None*
     * - **Default Value**: Secondary
     * - **API ID Path**: swiper.items[].button_variant
     * - **Documentation**: https://prismic.io/docs/field#select
     */
    button_variant: prismic.SelectField<
        'Secondary' | 'Base' | 'Ghost' | 'Quite' | 'Text',
        'filled'
    >;

    /**
     * Button Icon field in *Swiper → Items*
     *
     * - **Field Type**: Select
     * - **Placeholder**: *None*
     * - **API ID Path**: swiper.items[].button_icon
     * - **Documentation**: https://prismic.io/docs/field#select
     */
    button_icon: prismic.SelectField<
        '-' | 'Discover' | 'Globe' | 'Northeast' | 'Left' | 'Right'
    >;
}

/**
 * Collection variation for Swiper Slice
 *
 * - **API ID**: `collection`
 * - **Description**: Default
 * - **Documentation**: https://prismic.io/docs/slice
 */
export type SwiperSliceCollection = prismic.SharedSliceVariation<
    'collection',
    Simplify<SwiperSliceCollectionPrimary>,
    Simplify<SwiperSliceCollectionItem>
>;
// ...
export type SwiperSliceProductSkuFeature = prismic.SharedSliceVariation<
    'product_sku_feature',
    Simplify<SwiperSliceProductSkuFeaturePrimary>,
    Simplify<SwiperSliceProductSkuFeatureItem>
>;
// ...
export interface SwiperSliceCapsuleItem {
    /**
     * Desktop Image field in *Swiper → Items*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: swiper.items[].desktop_image
     * - **Documentation**: https://prismic.io/docs/field#image
     */
    desktop_image: prismic.ImageField<never>;

    /**
     * Mobile Image field in *Swiper → Items*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: swiper.items[].mobile_image
     * - **Documentation**: https://prismic.io/docs/field#image
     */
    mobile_image: prismic.ImageField<never>;

    /**
     * Button Label field in *Swiper → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: Call to Action
     * - **API ID Path**: swiper.items[].button_label
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    button_label: prismic.KeyTextField;

    /**
     * Link field in *Swiper → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: swiper.items[].link
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    link: prismic.KeyTextField;

    /**
     * Button Variant field in *Swiper → Items*
     *
     * - **Field Type**: Select
     * - **Placeholder**: *None*
     * - **Default Value**: Text
     * - **API ID Path**: swiper.items[].button_variant
     * - **Documentation**: https://prismic.io/docs/field#select
     */
    button_variant: prismic.SelectField<
        'Text' | 'Base' | 'Secondary' | 'Ghost' | 'Quite',
        'filled'
    >;

    /**
     * Button Icon field in *Swiper → Items*
     *
     * - **Field Type**: Select
     * - **Placeholder**: *None*
     * - **API ID Path**: swiper.items[].button_icon
     * - **Documentation**: https://prismic.io/docs/field#select
     */
    button_icon: prismic.SelectField<
        '-' | 'Northeast' | 'Right' | 'Left' | 'Discover' | 'Globe'
    >;
}

/**
 * Capsule variation for Swiper Slice
 *
 * - **API ID**: `capsule`
 * - **Description**: Default
 * - **Documentation**: https://prismic.io/docs/slice
 */
export type SwiperSliceCapsule = prismic.SharedSliceVariation<
    'capsule',
    Simplify<SwiperSliceCapsulePrimary>,
    Simplify<SwiperSliceCapsuleItem>
>;
// ...
export interface SwiperSliceProductFlatlaySingleCollectionItem {
    /**
     * Button Label field in *Swiper → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: swiper.items[].button_label
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    button_label: prismic.KeyTextField;

    /**
     * Button Link field in *Swiper → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: /collections/new-arrivals
     * - **API ID Path**: swiper.items[].button_link
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    button_link: prismic.KeyTextField;

    /**
     * Button Variant field in *Swiper → Items*
     *
     * - **Field Type**: Select
     * - **Placeholder**: *None*
     * - **API ID Path**: swiper.items[].button_variant
     * - **Documentation**: https://prismic.io/docs/field#select
     */
    button_variant: prismic.SelectField<
        'Base' | 'Secondary' | 'Ghost' | 'Text' | 'Quite'
    >;

    /**
     * Button Icon field in *Swiper → Items*
     *
     * - **Field Type**: Select
     * - **Placeholder**: *None*
     * - **API ID Path**: swiper.items[].button_icon
     * - **Documentation**: https://prismic.io/docs/field#select
     */
    button_icon: prismic.SelectField<
        '-' | 'Globe' | 'Discover' | 'Northeast' | 'Right' | 'Left'
    >;
}

/**
 * Product Flatlay Single Collection variation for Swiper Slice
 *
 * - **API ID**: `product_flatlay_single_collection`
 * - **Description**: Default
 * - **Documentation**: https://prismic.io/docs/slice
 */
export type SwiperSliceProductFlatlaySingleCollection =
    prismic.SharedSliceVariation<
        'product_flatlay_single_collection',
        Simplify<SwiperSliceProductFlatlaySingleCollectionPrimary>,
        Simplify<SwiperSliceProductFlatlaySingleCollectionItem>
    >;
// ...
export type SwiperSliceProductFlatlayTabbedCollections =
    prismic.SharedSliceVariation<
        'product_flatlay_tabbed_collections',
        Simplify<SwiperSliceProductFlatlayTabbedCollectionsPrimary>,
        Simplify<SwiperSliceProductFlatlayTabbedCollectionsItem>
    >;

/**
 * Slice variation for *Swiper*
 */
type SwiperSliceVariation =
    | SwiperSliceProductFeature
    | SwiperSliceProductShopify
    | SwiperSliceCollection
    | SwiperSliceProductSkuFeature
    | SwiperSliceCapsule
    | SwiperSliceProductFlatlaySingleCollection
    | SwiperSliceProductFlatlayTabbedCollections;

/**
 * Swiper Shared Slice
 *
 * - **API ID**: `swiper`
 * - **Description**: Swiper
 * - **Documentation**: https://prismic.io/docs/slice
 */
export type SwiperSlice = prismic.SharedSlice<'swiper', SwiperSliceVariation>;